+++ src/config/routes/index.js
@@ -1,7 +1,3 @@
-import administrator from './administratorRoutes';
 import main from './mainRoutes';
 
-export {
-    administrator,
-    main,
-};
+export { main };
